+++ src/main/java/no/gunbang/market/common/entity/BaseEntity.java
@@ -1,4 +1,4 @@
-package no.gunbang.market.common;
+package no.gunbang.market.common.entity;
 
 import jakarta.persistence.Column;
 import jakarta.persistence.EntityListeners;
