+++ src/main/java/no/gunbang/market/common/entity/ItemRepository.java
@@ -1,4 +1,4 @@
-package no.gunbang.market.common;
+package no.gunbang.market.common.entity;
 
 import org.springframework.data.jpa.repository.JpaRepository;
