+++ src/main/java/no/gunbang/market/domain/auction/cursor/DueDateCursorStrategy.java
@@ -4,7 +4,7 @@ import com.querydsl.core.types.Order;
 import com.querydsl.core.types.Predicate;
 import com.querydsl.core.types.dsl.Expressions;
 import java.time.LocalDateTime;
-import no.gunbang.market.common.CursorStrategy;
+import no.gunbang.market.common.query.CursorStrategy;
 import no.gunbang.market.domain.auction.entity.QAuction;
 
 public class DueDateCursorStrategy implements CursorStrategy<AuctionCursorValues> {
