+++ src/main/java/no/gunbang/market/domain/auction/entity/Auction.java
@@ -15,9 +15,9 @@ import java.time.LocalDateTime;
 import java.util.Objects;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
-import no.gunbang.market.common.BaseEntity;
-import no.gunbang.market.common.Item;
-import no.gunbang.market.common.Status;
+import no.gunbang.market.common.entity.BaseEntity;
+import no.gunbang.market.common.entity.Item;
+import no.gunbang.market.common.entity.Status;
 import no.gunbang.market.common.exception.CustomException;
 import no.gunbang.market.common.exception.ErrorCode;
 import no.gunbang.market.domain.user.entity.User;
