+++ src/main/java/no/gunbang/market/domain/auction/entity/Bid.java
@@ -13,7 +13,7 @@ import jakarta.persistence.Table;
 import java.time.LocalDateTime;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
-import no.gunbang.market.common.BaseEntity;
+import no.gunbang.market.common.entity.BaseEntity;
 import no.gunbang.market.common.exception.CustomException;
 import no.gunbang.market.common.exception.ErrorCode;
 import no.gunbang.market.domain.user.entity.User;
