+++ src/main/java/no/gunbang/market/domain/auction/repository/AuctionRepositoryImpl.java
@@ -8,9 +8,9 @@ import com.querydsl.jpa.impl.JPAQueryFactory;
 import java.time.LocalDateTime;
 import java.util.List;
 import lombok.RequiredArgsConstructor;
-import no.gunbang.market.common.CursorStrategy;
+import no.gunbang.market.common.query.CursorStrategy;
 import no.gunbang.market.common.QItem;
-import no.gunbang.market.common.Status;
+import no.gunbang.market.common.entity.Status;
 import no.gunbang.market.domain.auction.cursor.AuctionCursorValues;
 import no.gunbang.market.domain.auction.cursor.AuctionDefaultCursorStrategy;
 import no.gunbang.market.domain.auction.cursor.CurrentMaxPriceCursorStrategy;
