+++ src/main/java/no/gunbang/market/domain/market/dto/MarketResponseDto.java
@@ -4,9 +4,8 @@ package no.gunbang.market.domain.market.dto;
 import com.querydsl.core.annotations.QueryProjection;
 import lombok.Builder;
 import lombok.Getter;
-import no.gunbang.market.common.Status;
+import no.gunbang.market.common.entity.Status;
 import no.gunbang.market.domain.market.entity.Market;
-import no.gunbang.market.domain.user.dto.UserResponseDto;
 
 @Builder
 @Getter
