+++ src/main/java/no/gunbang/market/domain/market/entity/Market.java
@@ -11,13 +11,12 @@ import jakarta.persistence.Id;
 import jakarta.persistence.JoinColumn;
 import jakarta.persistence.ManyToOne;
 import jakarta.persistence.Table;
-import jakarta.persistence.Version;
 import java.util.Objects;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
-import no.gunbang.market.common.BaseEntity;
-import no.gunbang.market.common.Item;
-import no.gunbang.market.common.Status;
+import no.gunbang.market.common.entity.BaseEntity;
+import no.gunbang.market.common.entity.Item;
+import no.gunbang.market.common.entity.Status;
 import no.gunbang.market.common.exception.CustomException;
 import no.gunbang.market.common.exception.ErrorCode;
 import no.gunbang.market.domain.user.entity.User;
