+++ src/main/resources/META-INF/services/org.hibernate.boot.model.FunctionContributor
@@ -1,1 +1,1 @@
-no.gunbang.market.common.CustomFunctionContributor
+no.gunbang.market.common.query.CustomFunctionContributor
