+++ src/test/java/no/gunbang/market/AuctionServiceConcurrencyTest.java
@@ -9,8 +9,8 @@ import java.util.concurrent.ExecutorService;
 import java.util.concurrent.Executors;
 import java.util.concurrent.atomic.AtomicInteger;
 import lombok.extern.slf4j.Slf4j;
-import no.gunbang.market.common.Item;
-import no.gunbang.market.common.ItemRepository;
+import no.gunbang.market.common.entity.Item;
+import no.gunbang.market.common.entity.ItemRepository;
 import no.gunbang.market.common.exception.CustomException;
 import no.gunbang.market.domain.auction.dto.request.BidAuctionRequestDto;
 import no.gunbang.market.domain.auction.entity.Auction;
