+++ src/test/java/no/gunbang/market/RegisterMarketTest.java
@@ -6,10 +6,10 @@ import static org.mockito.Mockito.when;
 import java.util.Map;
 import java.util.Optional;
 import lombok.extern.slf4j.Slf4j;
-import no.gunbang.market.common.Inventory;
-import no.gunbang.market.common.InventoryRepository;
-import no.gunbang.market.common.Item;
-import no.gunbang.market.common.ItemRepository;
+import no.gunbang.market.common.entity.Inventory;
+import no.gunbang.market.common.entity.InventoryRepository;
+import no.gunbang.market.common.entity.Item;
+import no.gunbang.market.common.entity.ItemRepository;
 import no.gunbang.market.common.exception.CustomException;
 import no.gunbang.market.common.exception.ErrorCode;
 import no.gunbang.market.domain.market.dto.MarketRegistrationRequestDto;
