+++ src/test/java/no/gunbang/market/TradeMarketTest.java
@@ -11,10 +11,10 @@ import java.util.List;
 import java.util.Map;
 import java.util.Optional;
 import lombok.extern.slf4j.Slf4j;
-import no.gunbang.market.common.InventoryRepository;
-import no.gunbang.market.common.Item;
-import no.gunbang.market.common.ItemRepository;
-import no.gunbang.market.common.Status;
+import no.gunbang.market.common.entity.InventoryRepository;
+import no.gunbang.market.common.entity.Item;
+import no.gunbang.market.common.entity.ItemRepository;
+import no.gunbang.market.common.entity.Status;
 import no.gunbang.market.domain.market.dto.MarketTradeRequestDto;
 import no.gunbang.market.domain.market.dto.MarketTradeResponseDto;
 import no.gunbang.market.domain.market.entity.Market;
