+++ src/test/java/no/gunbang/market/unit/domain/auction/entity/AuctionTest.java
@@ -4,7 +4,7 @@ import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertThrows;
 
 import no.gunbang.market.TestData;
-import no.gunbang.market.common.Item;
+import no.gunbang.market.common.entity.Item;
 import no.gunbang.market.common.exception.CustomException;
 import no.gunbang.market.common.exception.ErrorCode;
 import no.gunbang.market.domain.auction.entity.Auction;
