+++ src/test/java/no/gunbang/market/unit/domain/auction/service/AuctionServiceTest.java
@@ -9,9 +9,9 @@ import static org.mockito.Mockito.when;
 
 import java.util.Optional;
 import no.gunbang.market.TestData;
-import no.gunbang.market.common.Item;
-import no.gunbang.market.common.ItemRepository;
-import no.gunbang.market.common.Status;
+import no.gunbang.market.common.entity.Item;
+import no.gunbang.market.common.entity.ItemRepository;
+import no.gunbang.market.common.entity.Status;
 import no.gunbang.market.domain.auction.AuctionScheduler;
 import no.gunbang.market.domain.auction.dto.request.AuctionRegistrationRequestDto;
 import no.gunbang.market.domain.auction.dto.request.BidAuctionRequestDto;
